Initial commit for new eegr implementation.
When regression is run on a 2D grid with channels on one axis and
time/samples on the other, we get a 2D grid of 'buckets' of diagnostic
and fit information: betas, confidence intervals, influence measures and
so on.

That 2D grid is stored as a NumPy array of special bucket datatype, so
we can index into it and get any diagnostic measure or fit information
we want. In addition, the grid is wrapped in a FitGrid object to allow
the user to first select channels like so:

    grid['channel22']['fit']['betas']

The plan is to let the user inspect the grid visually, by plotting
heatmaps of diagnostic information, stripcharts of betas, etc.