Add slicer validation in FitGrid.
The only way to index into a FitGrid on the outer level is by passing
either a channel (response variable) name, or a list of channel names.
All other 'slices' result in an error.

In the future, a wildcard for 'all channels' should be added to be used
as:

    fitgrid[:]

which would create appear as slice(None, None, None) from the inside.