Setup grid construction for multiprocessing.
This commit creates a setup of the grid construction that could be used
for parallelization. A number of issue are preventing use of
multiprocessing.

First: it looks like linear algebra libraries are multicore, and that
interferes with using processes. Although that can be restricted by
setting environment variables like OMP_NUM_THREADS, we prefer not to
mess with that.

Second: for some reason, processing using multiprocessing.Pool results
in GroupBy.apply errors, while serial processing of the same inputs
using the same processor function doesn't.