Introduce Epochs object, implement OLS.
The idea is to first construct an Epochs object from an epochs table
dataframe. Then, by calling methods like `lm` or `glm`, we can run
various models on the epochs, returning FitGrid objects. These can then
be queried for model diagnostic parameters, fit results, and so on.

OLS is supported, MLM and GLM are up next.