Improve io.epochs_from_hdf and add tests for io.
Previously reading from hdf assumed no index, but presence of EPOCH_ID
and TIME in the columns. Turns out, you can get a multiindexed dataframe
by reading hdf files using pandas.

So now we look for these variables in the index and the columns, and
throw an error if we don't find them.