Revert to having only EPOCH_ID in the internal epochs table index.
In addition, the internal epochs table is now a copy, not a reference to
the original table.

We require that snapshots have only EPOCH_ID in the index in order to
have single index series as attributes in fit objects after running
models. Adding TIME to internal epochs table index breaks snapshots, as
it refers to the same dataframe, but the index has changed. This was
discovered when plotting averages (snapshots.mean()) fails.