Add FitGrid saving and loading (grid.save and fitgrid.load_grid).
FitGrid objects can now be saved to files and loaded from files. This is
achieved by pickling a tuple containing grid._grid and
grid._epoch_index. The highest protocol available in pickle is used. For
Python 3.6 this means protocol 4. Loading is done by unpickling this
tuple and creating a new FitGrid object using the constructor.