Move models to models.py, add LMFitGrid, LMERFitGrid.
This is a major refactor of fitgrid.

Models are now top level functions, not class methods.

Before:

    grid = epochs.lm(RHS='blah')

Now:

    grid = fitgrid.lm(epochs, RHS='blah')

FitGrid was subclassed: LMFitGrid and LMERFitGrid were added. They only
hold proper model objects, so the following method call on an LMFitGrid
produces a regular FitGrid:

    influence = grid.get_influence()

In the above, grid is an LMFitGrid, while influence is a FitGrid.

run_model returns a regular FitGrid, _run_model (note the underscore)
returns a dataframe of objects and should be used to implement models
that return custom subclasses of FitGrid.