Add Feather file reader (fitgrid.epochs_from_feather).
This only works with most recent versions of pandas (>=0.24.0) and pyarrow
(>=0.12.0). At the time of this commit, this version of pandas is only
available in conda-forge under the rc label (version 0.24.0rc1), pyarrow
also available from conda-forge (version 0.12.0a0). Use of older
versions of either of these packages will results in ImportError.