Softened np.allclose tests to let LMER results differ slightly across versions
Added test constants FIT_ATOL, FIT_RTOL to tests/context.py

These are for np.isclose(actual, expected, atol=FIT_ATL, rtol=FIT_RTOL)
comparisons of actual computed values with gold-standard snapshots
of expected values.

Values that are out of tolerance are reported with a warning.

For fresh versions of conda r-lmer, r-matrix the values are in
tolerance.

For previous but not that old releases of the packages, a few
idiosyncratic values in test_utils_summary.py::test_summarize are out
of tolerance. Not by much but it is useful to see and worth knowing
about.

The tolerance values can be set loose or stricter as needed.

added gold-standard lm and lmer summary data files